Move `ControlFlowGraph` & `EmitterContext` under `IR`
