+++ ARMeilleure/CodeGen/RegisterAllocators/HybridAllocator.cs
@@ -1,5 +1,4 @@
 using ARMeilleure.IntermediateRepresentation;
-using ARMeilleure.Translation;
 using System;
 using System.Diagnostics;
 using System.Numerics;
