+++ ARMeilleure/CodeGen/X86/X86Optimizer.cs
@@ -1,6 +1,5 @@
 ﻿using ARMeilleure.CodeGen.Optimizations;
 using ARMeilleure.IntermediateRepresentation;
-using ARMeilleure.Translation;
 using System.Collections.Generic;
 using static ARMeilleure.IntermediateRepresentation.Operand.Factory;
 using static ARMeilleure.IntermediateRepresentation.Operation.Factory;
