+++ ARMeilleure/Diagnostics/IRDumper.cs
@@ -1,5 +1,4 @@
 using ARMeilleure.IntermediateRepresentation;
-using ARMeilleure.Translation;
 using System;
 using System.Collections.Generic;
 using System.Linq;
