+++ ARMeilleure/IntermediateRepresentation/ControlFlowGraph.cs
@@ -1,9 +1,8 @@
-using ARMeilleure.IntermediateRepresentation;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 
-namespace ARMeilleure.Translation
+namespace ARMeilleure.IntermediateRepresentation
 {
     class ControlFlowGraph
     {
@@ -14,7 +13,7 @@ namespace ARMeilleure.Translation
         public BasicBlock Entry { get; }
         public IntrusiveList<BasicBlock> Blocks { get; }
         public BasicBlock[] PostOrderBlocks => _postOrderBlocks;
-        public int[] PostOrderMap => _postOrderMap; 
+        public int[] PostOrderMap => _postOrderMap;
 
         public ControlFlowGraph(BasicBlock entry, IntrusiveList<BasicBlock> blocks, int localsCount)
         {
