+++ ARMeilleure/Signal/NativeSignalHandler.cs
@@ -1,6 +1,5 @@
 ﻿using ARMeilleure.CodeGen;
 using ARMeilleure.IntermediateRepresentation;
-using ARMeilleure.Translation;
 using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
